#1401 - Don't restart framework services after update in certain use-cases

+++ ovs/lib/update.py
@@ -51,17 +51,18 @@ class UpdateController(object):
     _update_file = '/etc/ready_for_upgrade'
     _update_ongoing_file = '/etc/update_ongoing'
 
-    framework_packages = {'arakoon', 'openvstorage'}
-    volumedriver_packages = {'alba', 'arakoon', 'volumedriver-no-dedup-base', 'volumedriver-no-dedup-server'}
-    all_core_packages = framework_packages.union(volumedriver_packages)
-    core_packages_with_binaries = {'alba', 'arakoon', 'volumedriver-no-dedup-server'}
+    _packages_core = {'framework': {'arakoon', 'openvstorage'},
+                      'storagedriver': {'alba', 'arakoon', 'volumedriver-no-dedup-base', 'volumedriver-no-dedup-server'}}
+    _packages_core_blocking = _packages_core['framework'].difference(_packages_core['storagedriver'])
+
+    packages_core_all = _packages_core['framework'].union(_packages_core['storagedriver'])
 
     #########
     # HOOKS #
     #########
     @staticmethod
     @add_hooks('update', 'get_package_info_multi')
-    def get_package_information_core(client, package_info):
+    def _get_package_information_core(client, package_info):
         """
         Called by GenericController.refresh_package_information() every hour
 
@@ -86,11 +87,11 @@ class UpdateController(object):
             if client.username != 'root':
                 raise RuntimeError('Only the "root" user can retrieve the package information')
 
-            binaries = PackageManager.get_binary_versions(client=client, package_names=UpdateController.core_packages_with_binaries)
-            installed = PackageManager.get_installed_versions(client=client, package_names=UpdateController.all_core_packages)
-            candidate = PackageManager.get_candidate_versions(client=client, package_names=UpdateController.all_core_packages)
-            if set(installed.keys()) != set(UpdateController.all_core_packages) or set(candidate.keys()) != set(UpdateController.all_core_packages):
-                raise RuntimeError('Failed to retrieve the installed and candidate versions for packages: {0}'.format(', '.join(UpdateController.all_core_packages)))
+            binaries = PackageManager.get_binary_versions(client=client, package_names=UpdateController._packages_core['storagedriver'])
+            installed = PackageManager.get_installed_versions(client=client, package_names=UpdateController.packages_core_all)
+            candidate = PackageManager.get_candidate_versions(client=client, package_names=UpdateController.packages_core_all)
+            if set(installed.keys()) != set(UpdateController.packages_core_all) or set(candidate.keys()) != set(UpdateController.packages_core_all):
+                raise RuntimeError('Failed to retrieve the installed and candidate versions for packages: {0}'.format(', '.join(UpdateController.packages_core_all)))
 
             # Retrieve Arakoon information
             framework_arakoons = []
@@ -162,7 +163,7 @@ class UpdateController(object):
                             elif version:
                                 running_version = version
 
-                            if package_name not in UpdateController.all_core_packages:
+                            if package_name not in UpdateController.packages_core_all:
                                 raise ValueError('Unknown package dependency found in {0}'.format(version_file))
                             if package_name not in binaries:
                                 raise RuntimeError('Binary version for package {0} was not retrieved'.format(package_name))
@@ -279,7 +280,7 @@ class UpdateController(object):
 
                 # Retrieve the actual update information
                 for package_name, package_info in storagerouter.package_information[key].iteritems():
-                    if package_name not in UpdateController.all_core_packages:
+                    if package_name not in UpdateController.packages_core_all:
                         continue  # Only gather information for the core packages
 
                     information[key]['services_post_update'].update(package_info.pop('services_to_restart'))
@@ -327,10 +328,10 @@ class UpdateController(object):
 
     @staticmethod
     @add_hooks('update', 'package_install_multi')
-    def package_install_core(client, package_info, components):
+    def _get_packages_to_install_core(client, package_info, components):
         """
         Update the core packages
-        :param client: Client on which to execute update the packages
+        :param client: Unused
         :type client: SSHClient
         :param package_info: Information about the packages (installed, candidate)
         :type package_info: dict
@@ -338,26 +339,22 @@ class UpdateController(object):
         :type components: list
         :return: None
         """
-        if 'framework' not in components and 'storagedriver' not in components:
-            return
-
+        _ = client  # Backwards compatibility (Was used until Fargo 2.7.11.1-1)
         packages_to_install = {}
-        for pkg_name, pkg_info in package_info.iteritems():
-            if pkg_name in UpdateController.all_core_packages:
-                packages_to_install[pkg_name] = pkg_info
-        if not packages_to_install:
-            return
-
-        UpdateController._logger.debug('{0}: Executing hook {1}'.format(client.ip, inspect.currentframe().f_code.co_name))
-        for pkg_name, pkg_info in packages_to_install.iteritems():
-            UpdateController._logger.debug('{0}: Updating core package {1} ({2} --> {3})'.format(client.ip, pkg_name, pkg_info['installed'], pkg_info['candidate']))
-            PackageManager.install(package_name=pkg_name, client=client)
-            UpdateController._logger.debug('{0}: Updated core package {1}'.format(client.ip, pkg_name))
-        UpdateController._logger.debug('{0}: Executed hook {1}'.format(client.ip, inspect.currentframe().f_code.co_name))
+        for component in components:
+            for pkg_name in UpdateController._packages_core.get(component, set()):
+                if pkg_name in package_info:
+                    pkg_info = package_info[pkg_name]
+                    if pkg_name in UpdateController._packages_core_blocking:
+                        pkg_info['blocking'] = True
+                    else:
+                        pkg_info['blocking'] = False
+                    packages_to_install[pkg_name] = pkg_info
+        return packages_to_install
 
     @staticmethod
     @add_hooks('update', 'post_update_multi')
-    def post_update_core(client, components):
+    def _post_update_core(client, components):
         """
         Execute functionality after the openvstorage core packages have been updated
         For framework:
@@ -386,7 +383,7 @@ class UpdateController(object):
             contents = local_client.file_read(filename='/opt/OpenvStorage/run/{0}'.format(version_file))
             for part in contents.split(';'):
                 packages.add(part.split('=')[0])
-            if packages.issubset(UpdateController.volumedriver_packages) and 'storagedriver' in components:
+            if packages.issubset(UpdateController._packages_core['storagedriver']) and 'storagedriver' in components:
                 service_name = version_file.replace('.remove', '').replace('.version', '')
                 UpdateController._logger.debug('{0}: Removing service {1}'.format(client.ip, service_name))
                 ServiceManager.stop_service(name=service_name, client=local_client)
@@ -542,6 +539,7 @@ class UpdateController(object):
         filemutex = file_mutex('system_update', wait=2)
         ssh_clients = []
         services_stop_start = set()
+        start_fwk_after_failure = True
         try:
             filemutex.acquire()
             UpdateController._logger.debug('+++ Starting update +++')
@@ -598,45 +596,53 @@ class UpdateController(object):
             if len(services_post_update) > 0:
                 UpdateController._logger.debug('Services which will be restarted after update: {0}'.format(', '.join(sorted(services_post_update))))
 
+            # Set migration values (Used when update is retried)
+            if 'framework' in components:
+                Configuration.set('/ovs/framework/update/migration', {'dal': True, 'extensions': True})
+
             # Stop services
             if UpdateController.change_services_state(services=services_stop_start,
                                                       ssh_clients=ssh_clients,
                                                       action='stop') is False:
                 raise Exception('Stopping all services on every node failed, cannot continue')
 
-            # Install packages
-            # First install packages on all StorageRouters individually
+            # Collect the functions to be executed before they get overwritten by updated packages, so on each the same functionality is executed
             package_install_multi_hooks = Toolbox.fetch_hooks('update', 'package_install_multi')
             package_install_single_hooks = Toolbox.fetch_hooks('update', 'package_install_single')
-            if packages_to_update:
-                failures = False
+
+            packages_to_install = {}
+            for function in package_install_multi_hooks:
+                packages_to_install.update(function(client=None, package_info=packages_to_update, components=components))
+
+            # Install each package on all StorageRouters
+            for pkg_name, pkg_info in packages_to_install.iteritems():
                 for client in ssh_clients:
-                    UpdateController._logger.debug('{0}: Installing packages'.format(client.ip))
-                    for function in package_install_multi_hooks:
-                        try:
-                            function(client=client, package_info=packages_to_update, components=components)
-                        except Exception as ex:
-                            UpdateController._logger.error('{0}: Package installation hook {1} failed with error: {2}'.format(client.ip, function.__name__, ex))
-                            failures = True
+                    UpdateController._logger.debug('{0}: Updating package {1} ({2} --> {3})'.format(client.ip, pkg_name, pkg_info['installed'], pkg_info['candidate']))
+                    try:
+                        PackageManager.install(package_name=pkg_name, client=client)
+                        UpdateController._logger.debug('{0}: Updated package {1}'.format(client.ip, pkg_name))
+                    except Exception as ex:
+                        UpdateController._logger.exception('{0}: Updating package failed. {1}'.format(client.ip, ex))
+                        if pkg_info.get('blocking', False) is True:
+                            start_fwk_after_failure = False
 
-                if set(components).difference({'framework', 'storagedriver'}):
-                    # Second install packages on all ALBA nodes
-                    for function in package_install_single_hooks:
-                        try:
-                            function(package_info=packages_to_update, components=components)
-                        except Exception as ex:
-                            UpdateController._logger.exception('Package installation hook {0} failed with error: {1}'.format(function.__name__, ex))
-                            failures = True
+            # Install packages on all ALBA nodes
+            if set(components).difference(set(UpdateController._packages_core.keys())):
+                for function in package_install_single_hooks:
+                    try:
+                        function(package_info=None, components=components)
+                    except Exception as ex:
+                        UpdateController._logger.exception('Package installation hook {0} failed with error: {1}'.format(function.__name__, ex))
 
-                if failures is True:
-                    raise Exception('Installing the packages failed on 1 or more nodes')
+            if start_fwk_after_failure is False:
+                raise Exception('Installing the packages failed on 1 or more nodes')
 
             # Remove update file
             for client in ssh_clients:
                 client.file_delete(UpdateController._update_file)
 
-            # Migrate code
-            if 'framework' in components:
+            # Migrate extensions
+            if 'framework' in components or (Configuration.exists('/ovs/framework/update/migration') and Configuration.get('/ovs/framework/update/migration|extensions') is True):
                 failures = []
                 for client in ssh_clients:
                     UpdateController._logger.debug('{0}: Verifying extensions code migration is required'.format(client.ip))
@@ -657,6 +663,7 @@ class UpdateController(object):
                         failures.append('{0}: {1}'.format(client.ip, str(ex)))
                 if len(failures) > 0:
                     raise Exception('Failed to run the extensions migrate code on all nodes. Errors found:\n\n{0}'.format('\n\n'.join(failures)))
+                Configuration.set('/ovs/framework/update/migration|extensions', False)
 
             # Start memcached
             if 'memcached' in services_stop_start:
@@ -665,8 +672,8 @@ class UpdateController(object):
                                                        ssh_clients=ssh_clients,
                                                        action='start')
 
-            # Migrate model
-            if 'framework' in components:
+            # Migrate DAL
+            if 'framework' in components or (Configuration.exists('/ovs/framework/update/migration') and Configuration.get('/ovs/framework/update/migration|dal') is True):
                 UpdateController._logger.debug('Verifying DAL code migration is required')
                 old_versions = PersistentFactory.get_client().get('ovs_model_version') if PersistentFactory.get_client().exists('ovs_model_version') else {}
 
@@ -677,6 +684,7 @@ class UpdateController(object):
                 new_versions = PersistentFactory.get_client().get('ovs_model_version') if PersistentFactory.get_client().exists('ovs_model_version') else {}
                 if old_versions != new_versions:
                     UpdateController._logger.debug('Finished DAL code migration. Old versions: {0} --> New versions: {1}'.format(old_versions, new_versions))
+                Configuration.set('/ovs/framework/update/migration|dal', False)
 
             # Post update actions
             for client in ssh_clients:
@@ -706,7 +714,7 @@ class UpdateController(object):
             UpdateController._logger.debug('Another update is currently in progress!')
         except Exception as ex:
             UpdateController._logger.exception('Error during update: {0}'.format(ex))
-            if len(ssh_clients) > 0:
+            if len(ssh_clients) > 0 and start_fwk_after_failure is True:
                 UpdateController.change_services_state(services=services_stop_start,
                                                        ssh_clients=ssh_clients,
                                                        action='start')
